Add separate bridge file
